Redirect submits to the new backend service
Redirect transaction submits to the new backend service. Its endpoint
takes the body in base64 format and neither takes nor returns the hash,
therefore the body needs to be reformatted and when checking whether
the transaction was submitted successfuly, the hash from signed
transaction is used instead of the hash from response. The old submitter
was not deleted yet, since it can serve as a fallback option.