bitbox02: lazily load bitbox02-api and split code out of main bundle
Since the bitbox02-api package is quite heavy, we only want to load it
when actually unlocking a BitBox02.

We follow the third approach described in
https://webpack.js.org/guides/code-splitting: using `await
import(...)` to load the package on demand, which also automatically
splits the code into a new file. The `webpackChunkName: "bitbox02-api"` describes the
name of the new file in dist/js/.

The tsconfig change was needed for code splitting to kick in,
according to:
- https://github.com/webpack/webpack/issues/5703#issuecomment-357512412
- https://github.com/webpack/webpack/issues/5703#issuecomment-462868924